Items and Levels
Levels contain an Array that hold the arrays for each level.
Tempted to turn the main array into a multidimensional array with 3
rows, tools, treasure, artefacts for easier legibility for the coders
and future games.
Temporary values have also been updated in the Data.xml files. Nothing
is in stone yet.