UHHH
A month's worth of Updates.
-Pretty much everything that should work does
-One bug right now is:
When clicking an item in the loot display only the Sigil ring is moved
into a bag. On top of that, if you didn't click the sigil ring the game
adds another to the bag. Then when you tap/click the sigil ring in the
bags, it adds a ring into the lootpile for EACH ring in the bags. And
keeps the rings in the bags.
-What is needed now are controls and inputs for mobile devices
-Updated UI to make it look good
.includes portraits and icons