Tweaks and Bugs
Loot icons and icons of items in the party's inventory aren't being
created properly.
From what I can tell they're just referencing the original object and
not creating a new one (clone).
This is likely the case with any icon we should be able to see
duplicates of, assumin there were others we needed to see.
I recomend checking out ExcavationFunctions.cs
PopulateBagWagonAndPileLists
Lines 261 - 272 are where icons are made from the player's items
If there's a problem there and we fix it it can be applied to the wagon
and loot pile too.