Fixes and Breaks
Re-did some code to allow for more easily read and better funcitonal
item Display.
What should be happening when you click an Item Icon:
Item is moved to "opposite" container. If in pile, move it to
bags/wagon. If in stash move to bags. If in bags move to pile/stash as
appropiate to situation. Etc.
What actually happens:
That happens when it comes to stash items, except the icons are
invisible. They work, but are invisible. Loot icons are invisible AND
don't work.
24 hours of coding, and this is where I've ended up.